backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
motor==3.3.1
python-multipart==0.0.6
celery==5.3.0
redis==5.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.4.2
pdfplumber==0.10.3
openai==1.3.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0